Correction for issue 83
The commit 97bd69e added lines

+        if ($row[0]['field']) {
+            return (int) $row[0]['field'];
+        } else if ($row[0]->field) {
+            return (int) $row[0]->field;
+        }

to correct the count when using different fetch modes. First, if $row[0] is stdClass, the [] operator cannot be used. Second, those conditions don't work when count is 0, because 0 in PHP is false. You should be checking is_array or is_object to know how to return the count.